:recycle: refactor: debug-tools -> for commands
