+++ packages/debug-tools/package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "clsx": "^1.1.1",
     "got": "^11.8.2",
+    "mincu-debug": "^2.1.1",
     "react-use": "^17.3.1",
     "serve-handler": "^6.1.3",
     "tough-cookie": "^4.0.0"
