+++ packages/debug-tools/src/server.ts
@@ -1,8 +1,7 @@
 import { createServer, IncomingMessage, ServerResponse } from 'http'
 import path from 'path'
 import serveHandler from 'serve-handler'
-import got from 'got'
-import { CookieJar } from "tough-cookie";
+import { fetch } from './api/fetch'
 
 const PUBLIC_PATH = path.resolve(__dirname, '../../output')
 export const DEV_TOOL_PORT = 23333
@@ -22,25 +21,10 @@ const cors = (req: IncomingMessage, res: ServerResponse) => {
 // Exported for mincu-debug
 export const startDevTool = () => {
   createServer(async (req, res) => {
+    cors(req, res)
+
     if (req.url?.includes('/api/fetch')) {
-      // reverse proxy
-      cors(req, res)
-      const params = new URLSearchParams(req.url.split('?')[1])
-      const target = params.get('target')
-      if (target) {
-        const cookieJar = new CookieJar();
-        try {
-          const response = await got(decodeURIComponent(target),
-            { timeout: 10000, cookieJar }
-          );
-          res.end(response.body);
-          return;
-        } catch (e: any) {
-          return res.end(e.message);
-        }
-      } else {
-        res.end('No target')
-      }
+      return fetch(req, res)
     } else {
       return serveHandler(req, res, {
         public: PUBLIC_PATH,
